requirements.txt
aiogram
python-dotenv  # If you're using a .env file
